Basic Registration functionality working
User can now register an account and then log in with it at the log in
page. Also validation on both client side and server side is working.

Correct usernames are 3-16 chars, letters, numbers, dashes, and
underscores.

Correct passwords are any regular printable ASCII, 8-256 chars. Obviously
overkill on password lenght but why should length be restricted, and of
courese they are stored hashed on server.

All this code is very hackish and really should be refactored and
redesigned considerbly, but it works for the most part.