Switch authMgmt with Express Sessions
The session management that was being used in the authMgmt module has been
switched to Express session management.

The authentication code has been brought over into server.js. The same
functionality is here as before but simplified with Express session
management.

The next thing that would be desirable is to move all the authentication
back to a module again and then use it like a regular express middleware.